added a chat app and chat to my fish site

+++ appConditionals/public/js/feedback.js
@@ -10,17 +10,17 @@ $(function(){
           }, updateFeedback);
         });
 
-    $('.feedback-delete').click(function(e){
-      console.log(e.target.id);
-      e.preventDefault();
-      $.ajax({
-        url: '/api/0',
-        type: 'DELETE',
-        success: function(result) {
-            // Do something with the result
-        }
-    });
-    })    
+        $('.feedback-messages').on('click', function(e) {
+          console.log(e)
+          if (e.target.className == 'glyphicon glyphicon-remove') {
+            console.log('hello');
+            $.ajax({
+              url: 'api/' + e.target.id,
+              type: 'DELETE',
+              success: updateFeedback
+            }); //ajax
+          } // the target is a delete button
+      }); //feedback messages
         
       
         function updateFeedback(data) {
@@ -28,7 +28,7 @@ $(function(){
          $.each(data,function(key, item) {
            output += '     <div class="feedback-item item-list media-list">';
            output += '       <div class="feedback-item media">';
-           output += '       <div class="media-left"><button class="feedback-delete btn btn-xs btn-danger"><span id="' + key + '" class="glyphicon glyphicon-remove"></span></button></div>';
+           output += '       <div class="media-left"><button class="feedback-delete btn btn-xs btn-danger"><span id="' + key + '" class="glyphicon glyphicon-remove">X</span></button></div>';
            output += '         <div class="feedback-info media-body">';
            output += '           <div class="feedback-head">';
            output += '             <div class="feedback-title">' + item.title + ' <small class="feedback-name label label-info">' + item.name + '</small></div>';
